+++ src/app/app-material.module.ts
@@ -7,13 +7,20 @@
 
 import { NgModule } from '@angular/core';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { CenteredMatProgressSpinnerComponent } from './components/centered-mat-spinner/centered-mat-spinner.component';
 
 const MATERIAL_MODULES = [
   MatProgressSpinnerModule
 ];
 
 @NgModule({
+  declarations: [
+    CenteredMatProgressSpinnerComponent
+  ],
   imports: MATERIAL_MODULES,
-  exports: MATERIAL_MODULES
+  exports: [
+    ...MATERIAL_MODULES,
+    CenteredMatProgressSpinnerComponent
+  ]
 })
 export class AppMaterialModule { }