+++ src/app/app.component.ts
@@ -5,17 +5,35 @@
  * https://opensource.org/licenses/MIT
  */
 
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { ActivationEnd, Router } from '@angular/router';
+import { merge, Observable, of } from 'rxjs';
+import { delay, filter, mapTo, startWith, take } from 'rxjs/operators';
 
 @Component({
   selector: 'app-root',
-  template: '<router-outlet></router-outlet>',
+  templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent {
+export class AppComponent
+  implements OnInit {
 
+  isLoading$: Observable<boolean>;
 
+  constructor(
+    private router: Router
+  ) { }
 
-  constructor() { }
+  ngOnInit(): void {
+    this.isLoading$ = merge(
+      of(true),
+      this.router.events.pipe(
+        filter(ev => ev instanceof ActivationEnd),
+        take(1),
+        delay(50),
+        mapTo(false)
+      )
+    );
+  }
 
 }